requirements.txt
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
Django==5.2.7
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
dotenv==0.9.9
filelock==3.19.1
fsspec==2025.9.0
gunicorn==23.0.0
hf-xet==1.1.10
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.4
packaging==25.0
python-dotenv==1.1.1
PyYAML==6.0.3
regex==2025.10.23
requests==2.32.5
safetensors==0.6.2
setuptools==70.2.0
sqlparse==0.5.3
sympy==1.14.0
tokenizers==0.22.1
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.5.0
